--- BSEC_1.4.8.0_Generic_Release/algo/normal_version/bin/RaspberryPi/PiThree_ArmV6/libalgobsec.a.Size.log
@@ -1,2 +1,0 @@
-   text	   data	    bss	    dec	    hex	filename
-  57648	      0	   1248	  58896	   e610	(TOTALS)
