--- BSEC_1.4.8.0_Generic_Release/bsec-library.c
@@ -1,401 +1,0 @@
-/* Copyright (C) 2017 alexh.name */
-/* I2C code by twartzek 2017 */
-/* argv[] code by TimothyBrown 2018 */
-/**
-  *  MIT License
-  *
-  *    Permission is hereby granted, free of charge, to any person obtaining a copy
-  *    of this software and associated documentation files (the "Software"), to deal
-  *    in the Software without restriction, including without limitation the rights
-  *    to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-  *    copies of the Software, and to permit persons to whom the Software is
-  *    furnished to do so, subject to the following conditions:
-  *
-  *    The above copyright notice and this permission notice shall be included in all
-  *    copies or substantial portions of the Software.
-  *
-  *    THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-  *    IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-  *    FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-  *    AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-  *    LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-  *    OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-  *    SOFTWARE.
-  */
-
-/*
- * Read the BME680 sensor with the BSEC library by running an endless loop in
- * the bsec_iot_loop() function under Linux.
- *
- */
-
-/* header files */
-
-#include <stdio.h>
-#include <stdlib.h>
-#include <time.h>
-#include <fcntl.h>
-#include <string.h>
-#include <unistd.h>
-#include <inttypes.h>
-#include <sys/ioctl.h>
-#include <sys/types.h>
-#include <sys/stat.h>
-#include <linux/i2c-dev.h>
-#include "bsec_datatypes.h"
-#include "bsec_integration.h"
-#include "bme680.h"
-
-/* definitions */
-int g_i2cFid; // I2C Linux device handle
-int i2c_address; // Changed from #define to argv[1].
-float temp_offset; // Changed from #define to argv[2].
-float sample_rate_mode; // Changed from #define to argv[3].
-char *filename_state = "bsec-library.state";
-char *filename_config = "bsec-library.config";
-
-/* functions */
-
-// open the Linux device
-void i2cOpen()
-{
-  g_i2cFid = open("/dev/i2c-1", O_RDWR);
-  if (g_i2cFid < 0) {
-    perror("i2cOpen");
-    exit(1);
-  }
-}
-
-// close the Linux device
-void i2cClose()
-{
-  close(g_i2cFid);
-}
-
-// set the I2C slave address for all subsequent I2C device transfers
-void i2cSetAddress(int address)
-{
-  if (ioctl(g_i2cFid, I2C_SLAVE, address) < 0) {
-    perror("i2cSetAddress");
-    exit(1);
-  }
-}
-
-/*
- * Write operation in either I2C or SPI
- *
- * param[in]        dev_addr        I2C or SPI device address
- * param[in]        reg_addr        register address
- * param[in]        reg_data_ptr    pointer to the data to be written
- * param[in]        data_len        number of bytes to be written
- *
- * return          result of the bus communication function
- */
-int8_t bus_write(uint8_t dev_addr, uint8_t reg_addr, uint8_t *reg_data_ptr,
-                 uint16_t data_len)
-{
-  int8_t rslt = 0; /* Return 0 for Success, non-zero for failure */
-
-  uint8_t reg[16];
-  reg[0]=reg_addr;
-
-  for (int i=1; i<data_len+1; i++)
-    reg[i] = reg_data_ptr[i-1];
-
-  if (write(g_i2cFid, reg, data_len+1) != data_len+1) {
-    perror("user_i2c_write");
-    rslt = 1;
-    exit(1);
-  }
-
-  return rslt;
-}
-
-/*
- * Read operation in either I2C or SPI
- *
- * param[in]        dev_addr        I2C or SPI device address
- * param[in]        reg_addr        register address
- * param[out]       reg_data_ptr    pointer to the memory to be used to store
- *                                  the read data
- * param[in]        data_len        number of bytes to be read
- *
- * return          result of the bus communication function
- */
-int8_t bus_read(uint8_t dev_addr, uint8_t reg_addr, uint8_t *reg_data_ptr,
-                uint16_t data_len)
-{
-  int8_t rslt = 0; /* Return 0 for Success, non-zero for failure */
-
-  uint8_t reg[1];
-  reg[0]=reg_addr;
-
-  if (write(g_i2cFid, reg, 1) != 1) {
-    perror("user_i2c_read_reg");
-    rslt = 1;
-  }
-
-  if (read(g_i2cFid, reg_data_ptr, data_len) != data_len) {
-    perror("user_i2c_read_data");
-    rslt = 1;
-  }
-
-  return rslt;
-}
-
-/*
- * System specific implementation of sleep function
- *
- * param[in]       t_ms    time in milliseconds
- *
- * return          none
- */
-void _sleep(uint32_t t_ms)
-{
-  struct timespec ts;
-  ts.tv_sec = 0;
-  /* mod because nsec must be in the range 0 to 999999999 */
-  ts.tv_nsec = (t_ms % 1000) * 1000000L;
-  nanosleep(&ts, NULL);
-}
-
-/*
- * Capture the system time in microseconds
- *
- * return          system_current_time    system timestamp in microseconds
- */
-int64_t get_timestamp_us()
-{
-  struct timespec spec;
-  //clock_gettime(CLOCK_REALTIME, &spec);
-  /* MONOTONIC in favor of REALTIME to avoid interference by time sync. */
-  clock_gettime(CLOCK_MONOTONIC, &spec);
-
-  int64_t system_current_time_ns = (int64_t)(spec.tv_sec) * (int64_t)1000000000
-                                   + (int64_t)(spec.tv_nsec);
-  int64_t system_current_time_us = system_current_time_ns / 1000;
-
-  return system_current_time_us;
-}
-
-/*
- * Handling of the ready outputs
- *
- * param[in]       timestamp       time in microseconds
- * param[in]       iaq             IAQ signal
- * param[in]       iaq_accuracy    accuracy of IAQ signal
- * param[in]       temperature     temperature signal
- * param[in]       humidity        humidity signal
- * param[in]       pressure        pressure signal
- * param[in]       raw_temperature raw temperature signal
- * param[in]       raw_humidity    raw humidity signal
- * param[in]       gas             raw gas sensor signal
- * param[in]       bsec_status     value returned by the bsec_do_steps() call
- *
- * return          none
- */
-void output_ready(int64_t timestamp, float iaq, uint8_t iaq_accuracy,
-                  float temperature, float humidity, float pressure,
-                  float raw_temperature, float raw_humidity, float gas,
-                  bsec_library_return_t bsec_status,
-                  float static_iaq, float co2_equivalent,
-                  float breath_voc_equivalent)
-{
-  //int64_t timestamp_s = timestamp / 1000000000;
-  ////int64_t timestamp_ms = timestamp / 1000;
-
-  //time_t t = timestamp_s;
-  /*
-   * timestamp for localtime only makes sense if get_timestamp_us() uses
-   * CLOCK_REALTIME
-   */
-  time_t t = time(NULL);
-  struct tm tm = *localtime(&t);
-
-  printf("{\"IAQ_Accuracy\": \"%d\"", iaq_accuracy);
-  printf(", \"IAQ\": \"%.2f\"", iaq);
-  printf(", \"Temperature\": \"%.2f\"", temperature);
-  printf(", \"Humidity\": \"%.2f\"", humidity);
-  printf(", \"Pressure\": \"%.2f\"", pressure / 100);
-  printf(", \"Gas\": \"%.0f\"", gas);
-  printf(", \"Status\": \"%d\"", bsec_status);
-// Added
-  printf(", \"IAQ_Static\": \"%.2f\"", static_iaq);
-  printf(", \"CO2e\": \"%.2f\"", co2_equivalent);
-  printf(", \"VOCe\": \"%.2f\"}", breath_voc_equivalent);
-
-  printf("\r\n");
-  fflush(stdout);
-}
-
-/*
- * Load binary file from non-volatile memory into buffer
- *
- * param[in,out]   state_buffer    buffer to hold the loaded data
- * param[in]       n_buffer        size of the allocated buffer
- * param[in]       filename        name of the file on the NVM
- * param[in]       offset          offset in bytes from where to start copying
- *                                  to buffer
- * return          number of bytes copied to buffer or zero on failure
- */
-uint32_t binary_load(uint8_t *b_buffer, uint32_t n_buffer, char *filename,
-                     uint32_t offset)
-{
-  int32_t copied_bytes = 0;
-  int8_t rslt = 0;
-
-  struct stat fileinfo;
-  rslt = stat(filename, &fileinfo);
-  if (rslt != 0) {
-    fprintf(stderr,"stat'ing binary file %s: ",filename);
-    perror("");
-    return 0;
-  }
-
-  uint32_t filesize = fileinfo.st_size - offset;
-
-  if (filesize > n_buffer) {
-    fprintf(stderr,"%s: %d > %d\n", "binary data bigger than buffer", filesize,
-            n_buffer);
-    return 0;
-  } else {
-    FILE *file_ptr;
-    file_ptr = fopen(filename,"rb");
-    if (!file_ptr) {
-      perror("fopen");
-      return 0;
-    }
-    fseek(file_ptr,offset,SEEK_SET);
-    copied_bytes = fread(b_buffer,sizeof(char),filesize,file_ptr);
-    if (copied_bytes == 0) {
-      fprintf(stderr,"%s empty\n",filename);
-    }
-    fclose(file_ptr);
-    return copied_bytes;
-  }
-}
-
-/*
- * Load previous library state from non-volatile memory
- *
- * param[in,out]   state_buffer    buffer to hold the loaded state string
- * param[in]       n_buffer        size of the allocated state buffer
- *
- * return          number of bytes copied to state_buffer or zero on failure
- */
-uint32_t state_load(uint8_t *state_buffer, uint32_t n_buffer)
-{
-  int32_t rslt = 0;
-  rslt = binary_load(state_buffer, n_buffer, filename_state, 0);
-  return rslt;
-}
-
-/*
- * Save library state to non-volatile memory
- *
- * param[in]       state_buffer    buffer holding the state to be stored
- * param[in]       length          length of the state string to be stored
- *
- * return          none
- */
-void state_save(const uint8_t *state_buffer, uint32_t length)
-{
-  FILE *state_w_ptr;
-  state_w_ptr = fopen(filename_state,"wb");
-  fwrite(state_buffer,length,1,state_w_ptr);
-  fclose(state_w_ptr);
-}
-
-/*
- * Load library config from non-volatile memory
- *
- * param[in,out]   config_buffer    buffer to hold the loaded state string
- * param[in]       n_buffer         size of the allocated state buffer
- *
- * return          number of bytes copied to config_buffer or zero on failure
- */
-uint32_t config_load(uint8_t *config_buffer, uint32_t n_buffer)
-{
-  int32_t rslt = 0;
-  /*
-   * Provided config file is 4 bytes larger than buffer.
-   * Apparently skipping the first 4 bytes works fine.
-   *
-   */
-  rslt = binary_load(config_buffer, n_buffer, filename_config, 4);
-  return rslt;
-}
-
-/* main */
-
-/*
- * Main function which configures BSEC library and then reads and processes
- * the data from sensor based on timer ticks
- *
- * return      result of the processing
- */
-int main(int argc, char *argv[])
-{
-  //putenv(DESTZONE); // Now taken care of in the Python controller.
-  if (argc == 4)
-    {
-      i2c_address = atoi (argv[1]);
-      if (i2c_address < 118 || i2c_address > 119)
-        {
-          printf("Error: '%s' is not a valid address for argument <i2c_address>.\nValid Options: 118|119\n", argv[1]);
-          return 1;
-        }
-      temp_offset = strtof (argv[2], NULL);
-      if (temp_offset > 10.0 || temp_offset < -10.0)
-        {
-          printf("Error: '%f' is outside of the valid range for argument <temperature_offset>.\nValid Range: 10.0 to -10.0\n", temp_offset);
-          return 1;
-        }
-      if (strcmp(argv[3], "LP") == 0)
-        {
-          sample_rate_mode = BSEC_SAMPLE_RATE_LP;
-        }
-      else if (strcmp(argv[3], "ULP") == 0)
-        {
-          sample_rate_mode = BSEC_SAMPLE_RATE_ULP;
-        }
-      else
-        {
-          printf("Error: '%s' isn't a valid option for argument <sample_rate_mode>.\nValid Options: LP|ULP\n", argv[3]);
-          return 1;
-        }
-    }
-  else
-    {
-      printf("Usage:\n");
-      printf("  %s <i2c_address> <temp_offset> <sample_rate_mode>\n", argv[0]);
-      printf("       i2c_address: 118|119\n       temp_offset: 10.0 to -10.0\n  sample_rate_mode: LP|ULP\n");
-      return 1;
-    }
-
-  i2cOpen();
-  i2cSetAddress(i2c_address);
-
-  return_values_init ret;
-
-  ret = bsec_iot_init(sample_rate_mode, temp_offset, bus_write, bus_read,
-                      _sleep, state_load, config_load);
-  if (ret.bme680_status) {
-    /* Could not intialize BME680 */
-    return (int)ret.bme680_status;
-  } else if (ret.bsec_status) {
-    /* Could not intialize BSEC library */
-    return (int)ret.bsec_status;
-  }
-
-  /* Call to endless loop function which reads and processes data based on
-   * sensor settings.
-   * State is saved every 10.000 samples, which means every 10.000 * 3 secs
-   * = 500 minutes (depending on the config).
-   *
-   */
-  bsec_iot_loop(_sleep, get_timestamp_us, output_ready, state_save, 10000);
-
-  i2cClose();
-  return 0;
-}
