--- BSEC_1.4.8.0_Generic_Release/config/generic_33v_300s_28d/bsec_serialized_configurations_iaq.h
@@ -1,4 +1,0 @@
-#include <stdint.h>
-
-extern const uint8_t bsec_config_iaq[454];
-
