--- BSEC_1.4.8.0_Generic_Release/readme_API.txt
@@ -1,4 +1,0 @@
-API link
-https://github.com/BoschSensortec/BME680_driver/releases
-
-please use version v3.5.10
