--- BSEC_1.4.8.0_Generic_Release/readme_BSEC_In_Arduino.txt
@@ -1,2 +1,0 @@
-please refer to github 
-https://github.com/BoschSensortec/BSEC-Arduino-library
